Fix PortableApps packaging.
Our Windows portable packaging environment has a lot of cruft which is
no longer relevant. We removed support for U3 packages and the method
we use to generate PortableApps packages has been deprecated for a while.

Create PortableApps packages using current file formats (AppInfo v3.0)
and tools. Generate the PA launcher using the PortableApps.com Launcher
generator. Copy files and directories from the top level instead of using
a manifest derived from the NSIS installer.

The manifest is a good idea, but we should create a central manifest
and use that to generate the NSIS and PortableApps packages instead of
trying to parse wireshark.nsi.

The new package still needs a bit of work but it installs and runs in
the current version of the PA Platform.

Remove the define for MAKENSIS_UNICODE. It doesn't look like we were
using it.

Start tearing down makefiles and scripts that we no longer use.

Ping-Bug: 4191
Change-Id: Ib7173eec887d0abf69bb176a1e3f943a5a63bee4
Reviewed-on: https://code.wireshark.org/review/3962
Reviewed-by: Gerald Combs <gerald@wireshark.org>