Make some variables size_t, as values of that width are assigned to it.
(I'd say "as values of that *type* are assigned to it", but the "len"
field of a GString is of type "gsize", and, for some unknown reason, the
GLib folks, rather than just using size_t, which is required by C89 to
be available, invented their own "gsize" type, defined as "long", which
happens to be narrower than "size_t" on LLP64 platforms such as 64-bit
Windows.  Sigh....)

This squelches some compiler warnings.

Change-Id: Ifca1300297b2e9b777fb2c426dda822ba157cfe5
Reviewed-on: https://code.wireshark.org/review/4014
Reviewed-by: Guy Harris <guy@alum.mit.edu>