Try to suppress MSVC++ unused parameter warnings.
I haven't found a way to with MSVC to mark parameters in the argument
list as unused.  MSVC doesn't give warnings about them in C code, but
does appear to give them with C++ code.  An answer to

	http://stackoverflow.com/questions/3020584/avoid-warning-unreferenced-formal-parameter

suggests not giving the formal parameter a name in C++.

Have a macro UNUSED_PARAMETER(), which takes as an argument a variable
name, and expands to nothing in C++ and to the variable name followed by
_U_ in C, and use that for some unused parameters.  If it works, we'll
use it for all of them.

Change-Id: I76107bed037f1f0d94615adb42234c9faf83b4db
Reviewed-on: https://code.wireshark.org/review/4016
Reviewed-by: Guy Harris <guy@alum.mit.edu>