Do the PCR and the original PCR the same way.
In both cases, make the extension a 16-bit variable, cast the result of
extracting the extension to guint16 to clarify that only the 9 bits
visible through the mask matter.

While we're at it, there's no need to use
"proto_tree_add_uint64_format_value() if the format is just the standard
format for a 64-bit unsigned integer.

Change-Id: I8f1f48595830d4672984f3797be1c9d994e64ea0
Reviewed-on: https://code.wireshark.org/review/4043
Reviewed-by: Guy Harris <guy@alum.mit.edu>