GTK+: Revert back to 2.24.14-1.1 on Win64.
As Pascal discovered in

https://www.wireshark.org/lists/wireshark-dev/201409/msg00045.html

Wireshark-gtk.exe crashes at startup on Windows 8.1 x64 when compiled
with the GTK+ 2.24.23-1.1 bundle and Visual C++ 2013. Revert to the
Win64 build to the prior GTK+ bundle, which works on my test system.

Update the release notes. Make sure we use a libintl-8.dll that's
compatible with GnuTLS.

Ideally we'd just upgrade or rebuild the GTK+ bundle but so far that
hasn't worked. Prior to this I tried:

Updating the gtk2 package at build.opensuse.org to 2.24.24. This fails with

[  187s] make[3]: Leaving directory `/home/abuild/rpmbuild/BUILD/gtk+-2.24.24/gtk'
[  187s] /usr/bin/gtk-update-icon-cache --force --ignore-theme-index		\
[  187s]    --include-image-data							\
[  187s]    --source builtin_icons stock-icons > gtkbuiltincache.h.tmp &&        \
[  187s] mv gtkbuiltincache.h.tmp gtkbuiltincache.h
[  187s] gtk-update-icon-cache: No theme index file.
[  187s] make[2]: *** [gtkbuiltincache.h] Error 1

Switching to the OBS GTK+ 3.14 package. It looks like a lot of our GTK+ code is
deprecated, including GtkAction and GtkAttachOptions.

Change-Id: I1548c84022f02895e5d424cd61e0fed7b57b2e75
Ping-Bug: 9914
Reviewed-on: https://code.wireshark.org/review/4379
Reviewed-by: Gerald Combs <gerald@wireshark.org>