Use the count-of-records fields to figure out how many records to read.
Fetch the count of records from one of the locations where it appears to
be, and, currently, require that it be equal to the count at the other
location where it appears to be; if they ever differ, we'll need the
file in order to reverse-engineer some more.

Fix the way we *write* .rf5 files - it turns out that we were

	1) not writing the full file size;

	2) not writing the packet count in the right location.

Detect files written by the old code, and get the packet count from the
right location for those files.

Change-Id: I7ce83afbc9dbbd300c81c96ef8f7785a0aeefa7a
Reviewed-on: https://code.wireshark.org/review/4608
Reviewed-by: Guy Harris <guy@alum.mit.edu>