IPv6: QuickStart option Nonce is read incorrectly
In Wireshark, the option's TTL field looks ok, but the [TTL diff] field points to the byte after the TTL field, which is part of the nonce field.
The nonce field is being read one byte after where it should be read and so the value being read is incorrect.
Moreover, the option data length says 6 but because of the extra byte being read 7 bytes are being read for this option data and the next option is read one byte later incorrectly so the options after the QuickStart option are being read incorrectly.

Found by boaz

Bug: 10575

Change-Id: Icd836b2a5c2f92cef7b6fea61957a85fca5e3352
Reviewed-on: https://code.wireshark.org/review/4815
Reviewed-by: Michael Mann <mmann78@netscape.net>
(cherry picked from commit e7b07ae73ec1af97170a5d4dd99f71eef6b4e84f)
Reviewed-on: https://code.wireshark.org/review/4825
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>