Fix a number of what appear to be errors.
Add checks for too-short length fields.

Increment the offset after some proto_tree_add_item() calls to skip past
the item.

Make some other length calculations use the start of the item to which
the length applies, not the start of the *list* of items.

Fix a double-digit field to be 2 bytes long (as the documentation says
it is).

Add a comment about a field that isn't always present but that's
specified in a tn5250_add_hf_items() list.

Fix DEFINE PITCH TABLE to match what the documentation appears to say it
is.

Change-Id: Ibcdc580045c68e8d0d8f35011dfe72b9c245e157
Reviewed-on: https://code.wireshark.org/review/4888
Reviewed-by: Guy Harris <guy@alum.mit.edu>