ncp2222: initialize buffer in build_expert_data
We do a bunch of conditional logic to fill in the buffer, and then were checking
strlen(buffer) to see if we'd actually filled it in or not, but if we hadn't
then the buffer was garbage and strlen(buffer) was throwing valgrind warnings.
Ensure this works as intended by setting the first byte of the buffer
unconditionally at the beginning, so strlen(buffer) returns 0 as expected in
that case.

Bug: 10628
Change-Id: I6f6c29fbbcb601b6ebccb3a4b312f0fa72c169ae
Reviewed-on: https://code.wireshark.org/review/4925
Reviewed-by: Evan Huus <eapache@gmail.com>
(cherry picked from commit b726d17816e0af213a554b6b4138e432ac002dd8)
Reviewed-on: https://code.wireshark.org/review/5353
Reviewed-by: Balint Reczey <balint@balintreczey.hu>