Handle the MCS index in Peek tagged files.
Use the "MCS index used" extended flag bit to indicate whether the "data
rate or MCS index" field is a data rate or MCS index.

Display the MCS index value if it's present.

(More to come - MCS indices, plus other information, should be used to
calculate the data rate for 11n and beyond.)

Get rid of the hdr_info_t structure while we're at it; just use local
variables for each of the fields.

Change-Id: I546f53a8ebd89078d5f23e1290557b97348aff38
Reviewed-on: https://code.wireshark.org/review/4988
Reviewed-by: Guy Harris <guy@alum.mit.edu>