Follow a suggestion from libtoolize:
libtoolize: Consider adding 'AC_CONFIG_MACRO_DIRS([./aclocal-fallback])' to
configure.ac

This is explained in the libtool manual:

"In the future other Autotools will automatically check the contents of
AC_CONFIG_MACRO_DIRS, but at the moment it is more portable to add the macro
directory to ACLOCAL_AMFLAGS in Makefile.am, which is where the tools currently
look. If libtoolize doesn't see AC_CONFIG_MACRO_DIRS, it too will
honour the first -I argument in ACLOCAL_AMFLAGS when
choosing a directory to store libtool configuration macros in. It is perfectly
sensible to use both AC_CONFIG_MACRO_DIRS and ACLOCAL_AMFLAGS, as long as they
are kept in synchronisation."

Change-Id: Ic203b0b5104ea3423cf8c0042b2278e82bcea332
Reviewed-on: https://code.wireshark.org/review/5113
Reviewed-by: Stephen Fisher <stephenfisher@centurylink.net>