dictionary: add missing TLV definition for WiMAX-Downlink-Granted-QoS
The type of the value of the Downlink-Granted-QoS is defined as tlv, but the
TLV definition has not given in the file.

This changeset adds the definition as per WiMAX Forum specification.
As per the spec, the TLV definition is same as the QoS-Descriptor TLV.

<Reference>

Section 5.4.2.29 QoS-Descriptor and 5.4.2.63 Downlink Granted-QoS in
WMF-T33-001-R010v05.

Change-Id: I72d97ed6f7532239cba62cc4cd1b6693fcd43123

dictionary: change type of WiMAX-Uplink-Granted-QoS VSA

This changeset changes the type of WiMAX-Uplink-Granted VSA to tlv and
provides the TLV definition as per WiMAX Forum Specification.

The TLV definition is same as the QoS-Descriptor VSA.

<Reference>

Section 5.4.2.30 Uplink Granted-QoS and 5.4.2.29 QoS-Descriptor in
WMF-T33-001-R010v05.

Change-Id: If6b505bc985c0fb41b19cf2d2fc7b4f887c72b7a
Reviewed-on: https://code.wireshark.org/review/5158
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>