Improve error message by including the (expected) class information
The code checks whether the tag AND class are matching with their expected values, but only the tag information is shown in the error message.
When the class is not matching, the error message shows the given and expected tags which are the same, confusing the user what is wrong with it.

Furthermore, the messages for SEQUENCE and SEQUENCE OF are made the same/similar.

Change-Id: I6c1d0968c903105ec7f6e0648d589a4263a8ca18
Reviewed-on: https://code.wireshark.org/review/5162
Reviewed-by: Stig Bjørlykke <stig@bjorlykke.org>