Do not force the output directory
For out of source builds, the output directory should not point to wireshark source locations.
By default it does point to the wireshark source directory (to stay compatible with the build in dissectors).
When the A2W_OUTPUT_DIR variable is set to "_EMPTY_" the output lcoation argument (-O) is not invoked.
Use A2W_FLAGS to set an output location for out of source builds (or use A2W_OUTPUT_DIR, which has the same result)

Change-Id: I48b7486fcda982f0dce57cde3beacb99f946abd9
Reviewed-on: https://code.wireshark.org/review/5254
Reviewed-by: Anders Broman <a.broman58@gmail.com>