Invoke asn2wrs.py using a custom command
A custom target is defined to be always invalid, so asn2wrs.py is always invoked.
A custom command is able to convert input file(s) into output file(s), the actual command is only invoked in the provided dependencies are changed.
As an advantage the dependencies are correct, so cmake is able to determine when the output file is required, so it can be build in time.

Change-Id: I84be5d408cfc87cf88d67b6073bea439febbe712
Reviewed-on: https://code.wireshark.org/review/5255
Reviewed-by: Anders Broman <a.broman58@gmail.com>