http: properly calculate end of chunked response
It was previously assumed that the remainder of a packet contains a
chunked-body response. This does not have be the case, and if the
assumption is violated, then the dissector would add multiple parts to a
single "De-chunked entity body".

This patch properly calculates the end of a chunked-body response,
taking the optional trailer-part into account and adjusting the size of
the chunked-body data as needed.

The CRLF in last-chunk that was previously dissected as "Chunk Boundary"
is the last CRLF that closes chunked-body, it is not part of last-chunk
(as it has no chunk-data to terminate).

A new header field is added for this trailer-part (RFC 7230 sec. 4.1).

Bug: 10707
Change-Id: Ifef1cc7dd0443edca4198eb1c27f58719f85fa9f
Reviewed-on: https://code.wireshark.org/review/5526
Reviewed-by: Peter Wu <peter@lekensteyn.nl>
Reviewed-by: Evan Huus <eapache@gmail.com>
Petri-Dish: Evan Huus <eapache@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>
(cherry picked from commit 6ee78f3be167807b9e2790b8f88f14ea1a4c9462)
(Conflicts due to context difference due to
e3a04bb392824781b8a3ef6ec98a37ef96c6c827 ("Ensure dissector data
parameter is used instead ..."); Drop proto_tree_add_subtree conversion
from 1.99 since it is missing in 1.12)
Reviewed-on: https://code.wireshark.org/review/7219