Squelch some "can't happen in reality" warnings.
(I guess newer versions of GCC/Clang know that dissect_eh_frame() is
never called with a segment_size of 0, so the loop is traversed at least
once.  NOTE: if it ever *is* called with a segment_size of 0, then
that's a genuine bug and needs to be fixed.)

Also, segment_size is used; no need to mark it as unused.

Change-Id: I63b7a580a853b55f22494de73b4c4e6f9a387647
Reviewed-on: https://code.wireshark.org/review/5591
Reviewed-by: Guy Harris <guy@alum.mit.edu>