Don't go past the end of the "extra data".
The "extra data", according to the RF5 API manual I have, is "hardware
parameters", so change some names.

Report an error if the variable parts of the configuration event record
run past the record length.

Report an error if the hardware parameters are too short.

For the purportedly DS0 hardware parameters, don't assume they'll be
long enough for the mask; they might not be.

Change-Id: Ib63d042e4ede32216fb474c4ecdba84db1387abc
Reviewed-on: https://code.wireshark.org/review/5638
Reviewed-by: Guy Harris <guy@alum.mit.edu>