Don't go past the end of the "extra data".
The "extra data", according to the RF5 API manual I have, is "hardware
parameters", so change some names.

Report an error if the variable parts of the configuration event record
run past the record length.

Report an error if the hardware parameters are too short.

For the purportedly DS0 hardware parameters, don't assume they'll be
long enough for the mask; they might not be.

Change-Id: I438f00645a2f7a24a4b2b02afc3bb0e8fe783f40
Reviewed-on: https://code.wireshark.org/review/5673
Reviewed-by: Guy Harris <guy@alum.mit.edu>