Don't put bitmasks under an FT_NONE.
proto_tree_add_bitmask() requires that the field under which it's
putting the bitmasks be an FT_INTn or FT_UINTn, so that it can determine
the length of the field.

Arguably, proto_tree_add_bitmask() should, instead, take a length
argument, just as other proto_tree_add_ routines do, and, arguably, we
should perhaps not even have FT_UINTn and FT_INTn, just FT_UINT and
FT_INT, with the display width for hex and octal (leading zeroes)
determined by the actual length of the field or something such as that,
or as part of the field that also contains the base.

But, even with that, we might want to require an FT_UINTn or FT_INTn
anyway, at least in cases where the value of the field as a whole is
interesting.

Change-Id: I4dff8fb1686a30b7d145c089dd1be7f96ecf23e0
Reviewed-on: https://code.wireshark.org/review/5680
Reviewed-by: Guy Harris <guy@alum.mit.edu>