Don't use WTAP_ERR_UNSUPPORTED for unwritable records.
WTAP_ERR_UNSUPPORTED is expected to include an err_info string;
WTAP_ERR_UNSUPPORTED_ENCAP isn't.  In the trunk, we use a new error code
for that, and supply and process the err_info string; here, we just use
WTAP_ERR_UNSUPPORTED_ENCAP, so we don't have to make as extensive changes.

Change-Id: Iaba47c409f9fd0d2edef3386d5d930dd2e370cab
Reviewed-on: https://code.wireshark.org/review/5825
Reviewed-by: Guy Harris <guy@alum.mit.edu>