Don't use WTAP_ERR_UNSUPPORTED for unwritable records.
WTAP_ERR_UNSUPPORTED is expected to include an err_info string;
WTAP_ERR_UNSUPPORTED_ENCAP isn't.  In the trunk, we use a new error code
for that, and supply and process the err_info string; here, we just use
WTAP_ERR_UNSUPPORTED_ENCAP, so we don't have to make as extensive changes.

Change-Id: I6f02238c178aff8e277bfc977c8b941d8e5ebea2
Reviewed-on: https://code.wireshark.org/review/5824
Reviewed-by: Guy Harris <guy@alum.mit.edu>